Stepping in bytecode
It should be possible to get sources, but saving them would be kind
of involved so I'm not so sure about that one. Also we really ought
to get arguments in native compiled functions.
