+++ src/llvmo/link_intrinsics.cc
@@ -58,6 +58,7 @@ extern "C" {
 #include <clasp/core/multipleValues.h>
 #include <clasp/core/compiler.h>
 #include <clasp/core/debugger.h>
+#include <clasp/core/step.h> // breakstep
 #include <clasp/core/backtrace.h>
 #include <clasp/llvmo/debugInfoExpose.h>
 #include <clasp/core/random.h>
@@ -537,53 +538,9 @@ extern "C" {
 void cc_set_breakstep() { my_thread->_Breakstep = true; }
 void cc_unset_breakstep() { my_thread->_Breakstep = false; }
 
-// RAII thing to toggle breakstep while respecting nonlocal exit.
-struct BreakstepToggle {
-  ThreadLocalState* mthread;
-  bool old_breakstep;
-  BreakstepToggle(ThreadLocalState* thread, bool new_breakstep) {
-    mthread = thread;
-    old_breakstep = thread->_Breakstep;
-    thread->_Breakstep = new_breakstep;
-  }
-  ~BreakstepToggle() { mthread->_Breakstep = old_breakstep; }
-};
-
 NOINLINE void cc_breakstep(core::T_O* source, void* frame) {
-  unlikely_if(my_thread->_Breakstep) {
-    void* bframe = my_thread->_BreakstepFrame;
-    // If bframe is NULL, we are doing step-into.
-    // Otherwise, we are doing step-over, and we need to check
-    // if we've returned yet. bframe is the frame step-over was initiated
-    // from, and lframe/frame is the caller frame.
-    // We have to check here because a function being stepped over may
-    // nonlocally exit past the caller, and in that situation we want to
-    // resume stepping.
-    // FIXME: We assume stack growth direction here.
-    if (!bframe || (frame >= bframe)) {
-      // Make sure we don't invoke the stepper recursively,
-      // but can do so again once we're out of the Lisp interaction.
-      BreakstepToggle tog(my_thread, false);
-      T_sp res = core::eval::funcall(core::_sym_breakstep, T_sp((gctools::Tagged)source));
-      if (res.fixnump()) {
-        switch (res.unsafe_fixnum()) {
-        case 0:
-          goto stop_stepping;
-        case 1:
-          my_thread->_BreakstepFrame = NULL;
-          return;
-        case 2:
-          my_thread->_BreakstepFrame = frame;
-          return;
-        }
-      }
-      SIMPLE_ERROR("BUG: Unknown return value from {}: {}", _rep_(core::_sym_breakstep), _rep_(res));
-    } else
-      return;
-  stop_stepping: // outside the scope of tog
-    my_thread->_Breakstep = false;
-    return;
-  }
+  unlikely_if(my_thread->_Breakstep)
+    breakstep(T_sp((gctools::Tagged)source), frame);
 }
 
 NOINLINE void cc_breakstep_after(void* frame) {
